Revised functionality and documentation updates
Updated documentation and help.
Added a new parameter to Format-Value to format as a number, so that
12345 will be formatted as 12,345.
Changed the Toggle value in Format-String to swap upper and lower case
as most people expect. Renamed what I had been using to Alternate. This
will always start with an upper case character.
Increased version numbers.